zmsg: xy precise calibrate: change data type to uint32_t

+++ include/zmsg/zmsg_fs_spec.hpp
@@ -24,7 +24,7 @@ public:
 	uint16_t motor_max_speed[motorId_t::NUM];
 
 	uint16_t motor_lzrz_fs_speed;
-	uint16_t motor_xy_precise_calibrate_speed;
+	uint32_t motor_xy_precise_calibrate_speed;
 
 	uint16_t motor_xy_steps_per_pixel;	// unit: step/pixel
